ingester: standardize tree names across origins
Add a yaml file for the community to collaborate on the
official KernelCI name/handle for each tree. This will
facilitate accessing results information in the dashboard
and kci-dev by just supplying the tree_name instead of
the full git_url.

This doesn't enforce having an entry for every git_url added
to KCIDB. Eg. some origins will use the tree name as the
differentiator for a tree that is specific to them, but they
submit as different "trees".

Signed-off-by: Gustavo Padovan <gustavo.padovan@collabora.com>